Refactor endpoint validation diagnostics helper
